Move rest of cli setup to cli.py. Make package callable via `python -m clonepool`
